64bit compatibility
del some un used files

+++ amffmpeg/libavutil/avutil.h
@@ -20,7 +20,7 @@
 
 #ifndef AVUTIL_AVUTIL_H
 #define AVUTIL_AVUTIL_H
-
+#include <stdint.h>
 /**
  * @file
  * external API header
@@ -109,7 +109,7 @@ enum AVMediaType {
 
 #define FF_QUALITY_SCALE FF_LAMBDA_SCALE //FIXME maybe remove
 
-#define AV_NOPTS_VALUE          INT64_C(0x8000000000000000)
+#define AV_NOPTS_VALUE          (int64_t)INT64_C(0x8000000000000000)
 #define AV_TIME_BASE            1000000
 #define AV_TIME_BASE_Q          (AVRational){1, AV_TIME_BASE}
